Refactor and modularize Tauri application commands
- Removed unused imports and optimized existing imports in `commands.rs`.
- Split command functions into separate modules for better organization:
  - `generate_session_id.rs` for session ID generation.
  - `get_available_models.rs` for fetching available AI models.
  - `get_chat_history.rs` for retrieving chat history.
  - `get_db_connection.rs` as a helper for database connection management.
  - `get_default_model.rs` for obtaining the default model setting.
  - `get_openai_api_key.rs` for accessing the OpenAI API key.
  - `set_openai_api_key.rs` for setting the OpenAI API key.
- Updated `lib.rs` to include new modules and reorganize imports.
- Improved functional style in message handling within the `chat_gpt` function.
- Enhanced error handling and logging for API requests and database operations.
